Make user id typed (add `UserId`)
Reason: when user id is just an integer, it's easy to accidentally
use a "random" integer instead of actual user id, typed APIs make code
less prone to some mistakes.

This commit also fixes an issue when a wrong integer type was used for
`user_id` in `Contact`.

Another option that this addition unlocks is adding methods/implementing
foreign trait for `UserId`.