Document the behavior of captions for media groups
As discussed in #987, the current behavior of captions regarding media
groups is documented in neither teloxide nor the Telegram API[0]. Users
of the official telegram clients may expect to be able to set what
appears to them as a "global" caption on media groups, when such a thing
does not actually exist in the API, and is just the result of how
official clients render media groups with only one captioned
`InputMedia`.

This commit changes the schema which auto-generates payload types, as
well as introduces the changes induced to the documentation of
`SendMediaGroup`.

Fix #987.

Links:
[0]: https://core.telegram.org/bots/api#sendmediagroup

Signed-off-by: Lymkwi <lymkwi@vulpinecitrus.info>